implemented new feature: definition files + test modification
changed java-parser (python and jason printer). Def files and file-defs are not part of the standard ast-depth first walker and handeld via defin and out of the visitor.
changed the interface of the savedefs method of the definition handeler to recursivly walk def-file trees. added a list of all visitied fileds to avoid circels
This code needs more than one test (TODO for the future)