first step of a refactoring/ addition of features to successfully eval more b-machines and solution states.
(1) the write-solution method (called if solution from proB is present before any parsing,typechecking or interpretation)
do not eval solution inside solution files. Now it only saves the ast-nodes of the solution files. This nodes will interpreted later inside the setup and exec init methods. This methods only perform property and invariant checking,
if the solution was compleate.
(2) bstate: solutions arry-code. This code was deleted as result of the new design.
(3) push of the constans and variables-names top level fram of the bstate is now done in setup and init-exec.
(4) the parsing of the ast is now side-effect free. (before this refactoring the solutions from prob where also used in this phase)
(5) fixed a wrong timeout arg to proB inside the team-tests (alstom modell)

one drawback of this refactoring is the performance, alstom test now need 15, instead of 5 minutes. The solutions are now
interpreted insed of been read.
The advantage is the enabeling of more complex solutions which need (limited) calculation. This was not possible before this post